update ubuntu.md to Alpine reality
So, our new esmero-php container is Alpine Linux, and internal Alpine www-data user has a different uid than the ubuntu one. This updates some  commands to make sure we don't end assigning a wrong uid to our Drupal folders and also adds some small explanations. Finally add a lot of sudos to everything for the prerequisites. People should know that but better so. @giancarlobi give this a try!