Updates Digital Object Solr Search Metadata Display on Deployment
This is just a tiny update.  Metadata shown on Solr Search Results was tied to a Display Mode setup to use Media Display Number 4. Since for this release we introduced a very simple MODS Display and that one is ingested as 4 it was being  used to display there, which really looks awefull. Lets deliver the basic Descriptive metadata Display for data and then document how people can. change it or make it better. For those who have this version running, its just a matter of doing a git pull and syncing back the configs. We need to document this too.