Make composer run without PHP client memory limits
This tries to address the problem that happens on very small machines (4 Gbytes on the host) where Docker runs out of memory first before composer has even the change to inform that (returning a "killed" message). If enough memory is around, with this change at least we will give Composer the chance to use all of it.