ISSUE-49: Fix wrong TIF configuration
# What is this?
This was found by Tom from CSHL. Java2D seems not be able to process TIFF files in the latest version (which we updated to on Beta2). This pull also enables GIF via IIIF. This is a config/bug-fix. 

Also odd. Since `processor.selection_strategy = ManualSelectionStrategy` if even converted to automatic would fail.. sad. But, could also be an issue with the strategy, maybe JAVA2D requires the download and can not stream?