Update field.field.node.digital_object_collection.field_descriptive_metadata.yml
Make field required and add a default value. This was removed on Beta2 and allowed people to ingest Collections without a SBF at all. This can not longer happen in RC1 since the initial webform form is displayed expandend but this is needed for existing collections to be resaved correctly. The SBF messages saying your Object had no SBF before will keep showing and that is OK.