New to me! Admin user needs to have Administrator Role Assigned Manually?
So, we all know that independently of the role User:1 gets all the benefits. But on D8 we could install via Drush using the Standard profile and use a config/sync folder. Now we can not and have to default to the minimal one. This (fun!) also removed the Administrator Role from User 1! If not, a LOT of what depends on the role (e.g webform elements / basicallt anything that DOES not check of User ==1 wil fail. Had a moment of panic with partial visible Webforms!!!! @alliomeria !)

So we have to add an extra drush command. @pcambra wonder if you can give me a hand tomorrow and kinda introduce us to some "strangenesses" of D9 so we can maybe explore better/newer ways of deployment via drush?